fix PVS warnings & errors, fix textbox rendering, unify todos
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@1370 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ pin3d.cpp
@@ -101,7 +101,7 @@ void EnvmapPrecalc(const DWORD* const __restrict envmap, const DWORD env_xres, c
 		for(unsigned int x = 0; x < rad_env_xres; ++x)
 		{
 			// trafo from envmap to normal direction
-			const float phi = (float)x/(float)rad_env_xres * (float)(2.0*M_PI) + M_PI;
+			const float phi = (float)x/(float)rad_env_xres * (float)(2.0*M_PI) + (float)M_PI;
 			const float theta = (float)y/(float)rad_env_yres * (float)M_PI;
 			const Vertex3Ds n(sinf(theta) * cosf(phi), sinf(theta) * sinf(phi), cosf(theta));
